Refactoring of the experimental data
